Create a standaone StyleSheet package
