+++ packages/react-native-web/package.json
@@ -16,12 +16,11 @@
     "debounce": "^1.1.0",
     "deep-assign": "^2.0.0",
     "fbjs": "^0.8.16",
-    "hyphenate-style-name": "^1.0.2",
     "inline-style-prefixer": "^3.0.8",
-    "normalize-css-color": "^1.0.2",
     "prop-types": "^15.6.0",
     "react-art": "^16.2.0",
-    "react-timer-mixin": "^0.13.3"
+    "react-timer-mixin": "^0.13.3",
+    "rnw-stylesheet": "^0.1.0"
   },
   "peerDependencies": {
     "react": "16.x.x",
