--- packages/react-native-web/src/apis/StyleSheet/registry.js
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) 2016-present, Nicolas Gallagher.
- * All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow
- */
-
-import StyleRegistry from './StyleRegistry';
-const registry = new StyleRegistry();
-export default registry;
