+++ packages/stylesheet/__tests__/StyleRegistry-test.js
