+++ packages/stylesheet/src/StyleManager.js
@@ -10,7 +10,7 @@
 
 import { canUseDOM } from 'fbjs/lib/ExecutionEnvironment';
 import generateCss from './generateCss';
-import hash from '../../vendor/hash';
+import hash from './vendor/hash';
 import staticCss from './staticCss';
 
 const emptyObject = {};
