+++ packages/stylesheet/src/getHairlineWidth.js
