+++ packages/stylesheet/src/lib/flattenArray/index.js
