+++ packages/stylesheet/src/lib/prefixStyles/static.js
