+++ packages/stylesheet/src/lib/processColor/__tests__/index-test.js
