+++ packages/stylesheet/src/lib/processColor/index.js
