+++ packages/stylesheet/src/staticCss.js
